+++ src/main/kotlin/io/kjson/JSONContext.kt
@@ -33,7 +33,8 @@ import kotlin.reflect.typeOf
 import io.kjson.pointer.JSONPointer
 
 /**
- * A class to hold the current serialization or deserialization context.
+ * A class to hold the current serialization or deserialization context, containing the [JSONConfig] and a [JSONPointer]
+ * representing the location currently being processed.
  *
  * @author  Peter Wall
  */
@@ -46,27 +47,44 @@ class JSONContext private constructor(
 
     constructor(config: JSONConfig, pointer: JSONPointer) : this(config, pointer.tokensAsArray())
 
+    constructor(pointer: JSONPointer) : this(JSONConfig.defaultConfig, pointer.tokensAsArray())
+
+    /** The [JSONPointer] represented by this `JSONContext`. */
     val pointer: JSONPointer
         get() = JSONPointer.from(pointerTokens)
 
+    /**
+     * Create a `JSONContext` representing the nominated property child of an object.
+     */
     fun child(token: String): JSONContext {
         val n = pointerTokens.size
         val tokens = Array(n + 1) { if (it < n) pointerTokens[it] else token }
         return JSONContext(config, tokens)
     }
 
+    /**
+     * Create a `JSONContext` representing the nominated array item child of an array.
+     */
     fun child(index: Int): JSONContext = child(index.toString())
 
-    fun serialize(obj: Any?): JSONValue? = JSONSerializer.serialize(obj, config) // TODO change JSONSerializer to take JSONContext
+    /**
+     * Serialize the given object to a [JSONValue], using this context (for use within a `toJSON` lambda).
+     */
+    fun serialize(obj: Any?): JSONValue? = JSONSerializer.serialize(obj, this)
 
     /**
      * Deserialize a [JSONValue] to the inferred type using this context.
      */
     inline fun <reified T> deserialize(json: JSONValue?): T = JSONDeserializer.deserialize(typeOf<T>(), json, this) as T
 
-    @Suppress("UNCHECKED_CAST")
-    fun <T> deserialize(type: KType, json: JSONValue?): T = JSONDeserializer.deserialize(type, json, this) as T
+    /**
+     * Deserialize a [JSONValue] to the specified [KType] using this context.
+     */
+    fun deserialize(type: KType, json: JSONValue?): Any? = JSONDeserializer.deserialize(type, json, this)
 
+    /**
+     * Deserialize a [JSONValue] to the specified [KClass] using this context.
+     */
     @Suppress("UNCHECKED_CAST")
     fun <T : Any> deserialize(resultClass: KClass<T>, json: JSONValue?): T =
             JSONDeserializer.deserialize(resultClass.starProjectedType, json, this) as T